banco_dados/requirements.txt
autopep8==1.5.7
boto3==1.18.16
botocore==1.21.16
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mysql-connector==2.2.9
psycopg2==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
